refactor: unify index manager
