+++ README.md
@@ -55,7 +55,7 @@ npm test
 
 ## Работа с `index.json`
 
-Утилиты из `tools/index_manager.js` помогают поддерживать актуальный список уроков и других файлов. Основные функции:
+Утилиты из `logic/index_manager.js` помогают поддерживать актуальный список уроков и других файлов. Основные функции:
 
 - `validateFilePathAgainstIndex(path)` \u2014 сверяет путь с данными из индекса и подсказывает корректное расположение файла;
 - `getLessonPath(num)` \u2014 возвращает путь урока по номеру, создавая запись вида `memory/lessons/lesson_XX.md`, если ее еще нет.
