+++ logic/index_manager.js
@@ -20,6 +20,14 @@ const indexPath = path.join(__dirname, '..', 'memory', 'index.json');
 let indexData = null;
 let validationReport = null;
 
+function loadIndexSync() {
+  try {
+    return sort_by_priority(index_tree.listAllEntries());
+  } catch {
+    return [];
+  }
+}
+
 function extractNumber(name) {
   const m = name.match(/(\d+)/);
   return m ? m[1].replace(/^0+/, '') : null;
@@ -105,6 +113,106 @@ function readLocalIndex() {
     console.warn('[indexManager] failed to read local index', e.message);
     return [];
   }
+}
+
+function getContextFiles() {
+  return loadIndexSync()
+    .filter(e => e && e.context_priority === 'high' && e.path)
+    .map(e => e.path);
+}
+
+function getLessonByNumber(num) {
+  const n = String(parseInt(num, 10)).padStart(2, '0');
+  return (
+    loadIndexSync().find(e => e.type === 'lesson' && e.path.includes(n)) || null
+  );
+}
+
+function getByPath(p) {
+  return loadIndexSync().find(e => e.path === p) || null;
+}
+
+function getByTag(tag) {
+  return loadIndexSync().filter(e => Array.isArray(e.tags) && e.tags.includes(tag));
+}
+
+function filterByStatus(status) {
+  return loadIndexSync().filter(e => e.status === status);
+}
+
+function filterByTags(tags) {
+  const arr = Array.isArray(tags) ? tags : [tags];
+  return loadIndexSync().filter(
+    e => Array.isArray(e.tags) && arr.every(t => e.tags.includes(t))
+  );
+}
+
+function filterByDate(field, days) {
+  const since = Date.now() - days * 24 * 60 * 60 * 1000;
+  return loadIndexSync().filter(e => {
+    const ts = Date.parse(e[field]);
+    return !isNaN(ts) && ts >= since;
+  });
+}
+
+function filterByCategory(category) {
+  return loadIndexSync().filter(e => e.category === category);
+}
+
+function searchByKeyword(query) {
+  const re = new RegExp(query, 'i');
+  return loadIndexSync().filter(e => {
+    const fields = [e.title, e.summary];
+    if (Array.isArray(e.tags)) fields.push(...e.tags);
+    return fields.some(v => v && re.test(String(v)));
+  });
+}
+
+function getContextFilesForKeywords(keywords = []) {
+  let entries = loadIndexSync();
+  if (keywords.length) {
+    const reList = keywords.map(k => new RegExp(k, 'i'));
+    entries = entries.filter(e => {
+      const fields = [e.title, e.summary];
+      if (Array.isArray(e.tags)) fields.push(...e.tags);
+      return reList.some(r => fields.some(v => v && r.test(String(v))));
+    });
+  }
+  return sort_by_priority(entries)
+    .filter(e => e && e.path)
+    .map(e => e.path);
+}
+
+async function sortIndexByPriority() {
+  if (!indexData) await loadIndex();
+  indexData = sort_by_priority(indexData);
+  await saveIndex();
+  return indexData;
+}
+
+function getNextLesson(num) {
+  const current = parseInt(num, 10);
+  const lessons = loadIndexSync()
+    .filter(e => e.type === 'lesson')
+    .map(e => ({ entry: e, n: parseInt((e.path.match(/(\d+)/) || [])[0], 10) }))
+    .sort((a, b) => a.n - b.n);
+  for (const { entry, n } of lessons) {
+    if (n > current) return entry;
+  }
+  return null;
+}
+
+async function updateMetadata(p, field, value) {
+  if (!indexData) await loadIndex();
+  const entry = indexData.find(e => e.path === p);
+  if (!entry) return null;
+  entry[field] = value;
+  await saveIndex();
+  return entry;
+}
+
+function validatePath(p) {
+  return !!loadIndexSync().find(e => e.path === p);
 }
 
 
@@ -314,4 +422,18 @@ module.exports = {
   addToIndex,
   moveFileAndUpdateIndex,
   getValidationReport,
+  getContextFiles,
+  getLessonByNumber,
+  getByPath,
+  getByTag,
+  filterByStatus,
+  filterByTags,
+  filterByDate,
+  filterByCategory,
+  searchByKeyword,
+  getContextFilesForKeywords,
+  sortIndexByPriority,
+  getNextLesson,
+  updateMetadata,
+  validatePath,
 };
